% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtts.utils.R
\name{align.idx}
\alias{align.idx}
\alias{align.idx,nanotime,nanotime,nanoduration,nanoduration-method}
\alias{align.idx,nanotime,nanotime,missing,missing-method}
\alias{align.idx,nanotime,nanotime,missing,nanoduration-method}
\alias{align.idx,nanotime,nanotime,nanoduration,missing-method}
\alias{align.idx,nanotime,nanotime,nanoperiod,nanoperiod-method}
\alias{align.idx,nanotime,nanotime,missing,nanoperiod-method}
\alias{align.idx,nanotime,nanotime,nanoperiod,missing-method}
\title{Get the index of the alignment of one vector onto another}
\usage{
align.idx(x, y, start, end, ...)

\S4method{align.idx}{nanotime,nanotime,nanoduration,nanoduration}(
  x,
  y,
  start = as.nanoduration(0),
  end = as.nanoduration(0),
  sopen = FALSE,
  eopen = TRUE
)

\S4method{align.idx}{nanotime,nanotime,missing,missing}(
  x,
  y,
  start = as.nanoduration(0),
  end = as.nanoduration(0),
  sopen = FALSE,
  eopen = TRUE
)

\S4method{align.idx}{nanotime,nanotime,missing,nanoduration}(
  x,
  y,
  start = as.nanoduration(0),
  end = as.nanoduration(0),
  sopen = FALSE,
  eopen = TRUE
)

\S4method{align.idx}{nanotime,nanotime,nanoduration,missing}(
  x,
  y,
  start = as.nanoduration(0),
  end = as.nanoduration(0),
  sopen = FALSE,
  eopen = TRUE
)

\S4method{align.idx}{nanotime,nanotime,nanoperiod,nanoperiod}(
  x,
  y,
  start = as.nanoperiod(0),
  end = as.nanoperiod(0),
  sopen = FALSE,
  eopen = TRUE,
  tz
)

\S4method{align.idx}{nanotime,nanotime,missing,nanoperiod}(
  x,
  y,
  start = as.nanoperiod(0),
  end = as.nanoperiod(0),
  sopen = FALSE,
  eopen = TRUE,
  tz
)

\S4method{align.idx}{nanotime,nanotime,nanoperiod,missing}(
  x,
  y,
  start = as.nanoperiod(0),
  end = as.nanoperiod(0),
  sopen = FALSE,
  eopen = TRUE,
  tz
)
}
\arguments{
\item{x}{the \code{nanotime} vector to align from}

\item{y}{the \code{nanotime} vector to align to}

\item{start}{scalar or vector of same length as \code{y} of type
\code{nanoduration}; \code{start} is added to each element in
\code{y} and it then defines the starting point of the
interval under consideration for the alignement on that
element of \code{y}}

\item{end}{scalar or vector of same length as \code{y} of type
\code{nanoduration}; \code{start} is added to each element in
\code{y} and it then defines the ending point of the interval
under consideration for the alignement on that element of
\code{y}}
}
\value{
a vector of indices of the same length as \code{y}; this
    vector indexes into \code{x} and represents the points in
    \code{x} that are aligned with the points in \code{y}
}
\description{
\code{align.idx} returns the index of the alignment of \code{x} on \code{y}
}
\details{
In order to perform the alignment, intervals are created around
each elements in \code{y} using \code{start} and \code{end}. For
each such interval, the closest element in \code{x} is chosen. If
no element in \code{x} falls in the interval, then NaN is
returned.
}
\examples{
\dontrun{
align.idx(nanotime(c(10:14, 17:19)), nanotime(11:20), as.nanoduration(-1), as.nanoduration(0))
## [1]  2  3  4  5  5 NA  6  7  8  8
}
}
