% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtts.utils.R
\name{align}
\alias{align}
\title{Align a \code{data.table} onto a \code{nanotime} vector}
\usage{
align(x, y, start = as.integer64(0), end = as.integer64(0), func = NULL)
}
\arguments{
\item{x}{the \code{data.table} time-series to align from}

\item{y}{the \code{nanotime} vector to align to}

\item{start}{scalar or vector of same length as \code{y} of type
\code{integer64}; \code{start} is added to each element in
\code{y} and it then defines the starting point of the
interval under consideration for the alignement on that
element of \code{y}}

\item{end}{scalar or vector of same length as \code{y} of type
\code{integer64}; \code{start} is added to each element in
\code{y} and it then defines the ending point of the interval
under consideration for the alignement on that element of
\code{y}}

\item{func}{a function taking one argument and which provides an
arbitrary aggregation of its argument; if \code{NULL} then a
function which takes the closest observation is used.}
}
\value{
a \code{data.table} time-series of the same length as
    \code{y}; this is a subset of \code{x} with the
    \code{nanotime} index of \code{y}
}
\description{
\code{align} returns the subset of \code{data.table} \code{x} that
aligns on the temporal vector \code{y}
}
\details{
For each element in \code{y}, intervals are created around this
element with \code{start} and \code{end}. All the elements of
\code{x} that fall within this interval are given as argument to
the function \code{func}. The function \code{func} show reduce
this \code{data.frame} to one unique row that will be associated
with the \code{nanotime} value in \code{y}.
}
\examples{
\dontrun{
y <- nanotime((1:10)*1e9)
x <- data.table(index=nanotime((1:10)*1e9), data=1:10)
align(x, y, as.integer64(-1e9), as.integer64(1e9), colMeans)
}
}
