% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtts.utils.R
\name{grid.align,data.table,nanoduration-method}
\alias{grid.align,data.table,nanoduration-method}
\alias{grid.align,data.table,nanoperiod-method}
\title{Align a \code{data.table} onto a \code{nanotime} vector grid}
\usage{
\S4method{grid.align}{data.table,nanoduration}(
  x,
  by,
  func,
  ival = by,
  start = x[[1]][1] + by,
  end = tail(x[[1]], 1)
)

\S4method{grid.align}{data.table,nanoperiod}(
  x,
  by,
  func,
  ival = by,
  start = plus(x[[1]][1], by, tz),
  end = tail(x[[1]], 1),
  tz
)
}
\arguments{
\item{x}{the \code{data.table} time-series to align from}

\item{by}{interval specified in nanoseconds}

\item{func}{a function taking one argument and which provides an
arbitrary aggregation of its argument; if \code{NULL} then a
function which takes the closest observation is used.}

\item{start}{scalar \code{nanotime} defining the start of the
grid; by default the first element of \code{x} is taken.}

\item{end}{scalar \code{nanotime} defining the end of the grid; by
default the last element of \code{x} is taken.}
}
\value{
a \code{data.table} time-series of the same length as
    \code{y} with the aggregations computed by \code{func}
}
\description{
\code{grid.align} returns the subset of \code{data.table} \code{x}
that aligns on the grid defined by \code{by}, \code{start} and
\code{end}
}
\details{
A grid defined by the parameter \code{by}, \code{start} and
\code{end} is created. The function then does a standard alignment
of \code{x} onto this grid (see the \code{align} function)
}
\examples{
\dontrun{
one_second <- 1e9
x <- data.table(index=nanotime(cumsum(sin(seq(0.001, pi, 0.001)) * one_second)))
x <- x[, V2 := 1:nrow(x)]
setkey(x, index)
grid.align(x, as.nanoduration("00:01:00"), sum)
}
}
