% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtts.utils.R
\docType{methods}
\name{frequency,data.table-method}
\alias{frequency,data.table-method}
\title{Return the number of observations per interval}
\usage{
\S4method{frequency}{data.table}(x, by, ival = by, start = x[[1]][1],
  end = tail(x[[1]], 1))
}
\arguments{
\item{x}{the \code{data.table} time-series for which to calculate
the frequency}

\item{by}{interval specified in nanoseconds}

\item{start}{scalar \code{nanotime} defining the start of the
grid; by default the first element of \code{x} is taken.}

\item{end}{scalar \code{nanotime} defining the end of the grid; by
default the last element of \code{x} is taken.}
}
\value{
a \code{data.table} time-series with the number of
    observations in \code{x} that fall withing the intervals
    defined by \code{by}
}
\description{
\code{frequency} returns the number of observations in
\code{data.table} \code{x} when subdivided in the specified
interval
}
\details{
The interval specified in \code{by} is used to subdivide
\code{x}. The number of observations in \code{x} is then counted
for each interval, and the results are assigned to the end of each
interval.
}
\examples{
\dontrun{
one_second <- 1e9
one_minute <- 60 * one_second
x <- data.table(index=nanotime(cumsum(sin(seq(0.001, pi, 0.001)) * one_second)), 1)
frequency(x, as.integer64(one_minute))
}
}
